acp, cloud, ort, copilot, cli, examples: clear remaining pedantic lints
ort: lift the forward pass into run_hidden_states, which returns an owned
array, so the session lock is held for exactly the inference and not for
pooling; derive Mean pooling's divisor from the same pass that sums the
hidden states, which removes a lossy u32-to-f32 cast and keeps the count
in step with what was actually summed.

cli: box the four agent-sized futures rather than leaving 90 KB frames on
main's stack; carry a bash hook's exit code as the i64 the wire gives us
instead of truncating it to i32.

examples: propagate tool registration errors instead of dropping them,
drop the spawned memory-write handle explicitly, and delete the unused
format_memories left over from an earlier design.

acp, cloud, copilot: fold identical match arms, drop an async with no
await, replace a hand-written future with async fn, and document the
error conditions the pedantic lints asked for.

+++ cli/src/hook.rs
@@ -122,7 +122,7 @@ fn parse_bash_args(arguments: &str) -> Option<(String, Option<String>)> {
 struct BashOutput {
     stdout: Option<String>,
     stderr: Option<String>,
-    exit_code: i32,
+    exit_code: i64,
     task_id: Option<String>,
     status: Option<String>,
 }
@@ -152,7 +152,7 @@ fn parse_bash_result(result: &str) -> Option<BashOutput> {
     let exit_code = parsed
         .get("exit_code")
         .and_then(serde_json::Value::as_i64)
-        .unwrap_or(0) as i32;
+        .unwrap_or(0);
     let task_id = parsed
         .get("task_id")
         .and_then(|v| v.as_str())
